pyproject.toml
[project]
name = "tp2-exploration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "altair>=6.0.0",
    "datasets>=4.4.1",
    "duckdb>=1.4.3",
    "gradio>=5.23.1",
    "httpx>=0.28.1",
    "huggingface-hub>=0.36.0",
    "litellm>=1.80.10",
    "notebook>=7.5.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
